JSApp refactoring; set up authnz controller to be able to serve a full client on the auth callback.
